create utils for dataframe funcs
