+++ listenbrainz_spark/tests/__init__.py
@@ -6,6 +6,7 @@ from datetime import datetime
 
 import listenbrainz_spark
 import listenbrainz_spark.utils.mapping as mapping_utils
+from listenbrainz_spark.recommendations import dataframe_utils
 from listenbrainz_spark import hdfs_connection, utils, config, schema
 from listenbrainz_spark.stats.utils import get_latest_listen_ts
 from listenbrainz_spark.recommendations import create_dataframes
@@ -200,9 +201,9 @@ class SparkTestCase(unittest.TestCase):
 
     @classmethod
     def upload_test_mapped_listens_to_hdfs(cls, listens_path, mapping_path, mapped_listens_path):
-        partial_listen_df = create_dataframes.get_listens_for_training_model_window(cls.date, cls.date, {}, listens_path)
+        partial_listen_df = dataframe_utils.get_listens_for_training_model_window(cls.date, cls.date, listens_path)
         df = utils.read_files_from_HDFS(mapping_path)
         mapping_df = mapping_utils.get_unique_rows_from_mapping(df)
 
-        mapped_listens = create_dataframes.get_mapped_artist_and_recording_mbids(partial_listen_df, mapping_df)
+        mapped_listens = dataframe_utils.get_mapped_artist_and_recording_mbids(partial_listen_df, mapping_df)
         utils.save_parquet(mapped_listens, mapped_listens_path)
